Stripe Signup Link for Instructors
Addition of a screen for instructors who haven't registered for a Stripe account for payment purposes.

This screen generates a unique Stripe Account Link for them to sign up as a Stripe subaccount to allow for payments to their own card of choice.

This screen allows for these instructors to receive a link, copy it and ultimately create an account.

Co-Authored-By: Julie Trinh <67767875+julietrinh1@users.noreply.github.com>