Backend logic to keep track of relevant sellers in a transaction
Backend and frontend adjustments to cascade relevant information to other screens in the payment process.

This was needed because the processing of payments need more than just the total price, since we need to know the pricing splits between all the relevant sellers in a transactions.

This commit simplifies the transaction and creates an array with the specific money to be distributed per seller/tutor.

Co-Authored-By: Julie Trinh <67767875+julietrinh1@users.noreply.github.com>